requirements.txt
allure-pytest==2.7.0
allure-python-commons==2.7.0
apipkg==1.5
appdirs==1.4.3
astroid==2.5
atomicwrites==1.4.0
attrs==20.3.0
CacheControl==0.12.6
cachetools==4.2.1
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
coverage==5.4
distlib==0.3.0
distro==1.4.0
execnet==1.8.0
google-auth==1.27.0
grpcio==1.26.0
grpcio-tools==1.26.0
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
isort==4.3.21
joblib==1.0.1
kubernetes==10.0.1
lazy-object-proxy==1.5.2
lockfile==0.12.2
mccabe==0.6.1
more-itertools==8.7.0
msgpack==0.6.2
numpy==1.18.1
oauthlib==3.1.0
packaging==20.3
pep517==0.8.2
pluggy==0.13.1
progress==1.5
protobuf==3.15.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.5.0
pymilvus-distributed==0.0.36
pymilvus-test==0.5.4
pyparsing==2.4.6
pytest==4.5.0
pytest-cov==2.8.1
pytest-forked==1.3.0
pytest-json-report==1.2.4
pytest-level==0.1.1
pytest-metadata==1.11.0
pytest-print==0.1.2
pytest-repeat==0.8.0
pytest-tags==1.7.2
pytest-timeout==1.3.3
pytest-xdist==1.23.2
python-dateutil==2.8.1
pytoml==0.1.21
PyYAML==5.4.1
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7.2
scikit-learn==0.24.1
scipy==1.6.1
six==1.14.0
sklearn==0.0
threadpoolctl==2.1.0
toml==0.10.2
ujson==4.0.2
urllib3==1.25.8
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
wrapt==1.12.1
